coremap allocation + TLB dirty bit

+++ os161-base-2.0.3/kern/include/vm.h
@@ -62,7 +62,7 @@ void vm_tlbshootdown(const struct tlbshootdown *);
 
 #if OPT_PAGING
 void vm_shutdown(void);
-paddr_t getppages(unsigned long npages);
+paddr_t getppages(unsigned npages);
 #endif
